# docs/openapi.json
{
 "components": {
  "schemas": {
...
   "EventsPage": {
    "properties": {
     "has_more": {
      "title": "Has More",
      "type": "boolean"
     },
     "index_updated_utc": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Index Updated Utc"
     },
     "items": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Items",
      "type": "array"
     },
     "limit": {
      "title": "Limit",
      "type": "integer"
     },
     "offset": {
      "title": "Offset",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "limit",
     "offset",
     "has_more"
    ],
    "title": "EventsPage",
    "type": "object"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "OceanKind",
  "version": "2.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/sites/{site_id}/acoustic": {
   "get": {
    "operationId": "acoustic_api_sites__site_id__acoustic_get",
    "parameters": [
     {
      "in": "path",
      "name": "site_id",
      "required": true,
      "schema": {
       "title": "Site Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Acoustic",
    "tags": [
     "data"
    ]
   }
  },
  "/api/sites/{site_id}/clips/{path}": {
   "get": {
    "description": "Audio proxied through the API, so the container stays private (R-5.4).",
    "operationId": "clip_api_sites__site_id__clips__path__get",
    "parameters": [
     {
      "in": "path",
      "name": "site_id",
      "required": true,
      "schema": {
       "title": "Site Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "path",
      "required": true,
      "schema": {
       "title": "Path",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Clip",
    "tags": [
     "data"
    ]
   }
  },
  "/api/sites/{site_id}/events": {
   "get": {
    "description": "Paginated and filtered. The browser gets a page, never the history (R-5.1).\n\nServed from the derived index, so this costs one query and no reads against\nobject storage regardless of the window (D-021, R-12.1).",
    "operationId": "events_api_sites__site_id__events_get",
    "parameters": [
     {
      "in": "path",
      "name": "site_id",
      "required": true,
      "schema": {
       "title": "Site Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "since",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date-time",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Since"
      }
     },
     {
      "in": "query",
      "name": "until",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date-time",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Until"
      }
     },
     {
      "in": "query",
      "name": "event_type",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "pattern": "^(vessel|blast|unknown)$",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Event Type"
      }
     },
     {
      "in": "query",
      "name": "min_score",
      "required": false,
      "schema": {
       "default": 0.0,
       "maximum": 1.0,
       "minimum": 0.0,
       "title": "Min Score",
       "type": "number"
      }
     },
     {
      "in": "query",
      "name": "include_suppressed",
      "required": false,
      "schema": {
       "default": true,
       "title": "Include Suppressed",
       "type": "boolean"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 50,
       "maximum": 500,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "minimum": 0,
       "title": "Offset",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventsPage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Events",
    "tags": [
     "data"
    ]
   }
  },
  "/api/sites/{site_id}/ocean": {
   "get": {
    "operationId": "ocean_api_sites__site_id__ocean_get",
    "parameters": [
     {
      "in": "path",
      "name": "site_id",
      "required": true,
      "schema": {
       "title": "Site Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Ocean",
    "tags": [
     "data"
    ]
   }
  },
  "/api/sites/{site_id}/power": {
   "get": {
    "operationId": "power_api_sites__site_id__power_get",
    "parameters": [
     {
      "in": "path",
      "name": "site_id",
      "required": true,
      "schema": {
       "title": "Site Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Power",
    "tags": [
     "data"
    ]
   }
  },
  "/api/sites/{site_id}/status": {
   "get": {
    "operationId": "status__api_sites__site_id__status_get",
    "parameters": [
     {
      "in": "path",
      "name": "site_id",
      "required": true,
      "schema": {
       "title": "Site Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Status ",
    "tags": [
     "data"
    ]
   }
  }
 }
}